chore: rename dialogs -> dialog, messages -> message
