+++ packages/core/src/v2/context.ts
@@ -2,9 +2,9 @@ import type { TelegramClient } from 'telegram'
 import type { ClientInstanceEventFromCore, ClientInstanceEventToCore } from './instance'
 import type { SessionEventFromCore, SessionEventToCore } from './services'
 import type { ConnectionEventFromCore, ConnectionEventToCore } from './services/connection'
-import type { DialogEventFromCore, DialogEventToCore } from './services/dialogs'
+import type { DialogEventFromCore, DialogEventToCore } from './services/dialog'
 import type { EntityEventFromCore, EntityEventToCore } from './services/entity'
-import type { MessageEventFromCore, MessageEventToCore } from './services/messages'
+import type { MessageEventFromCore, MessageEventToCore } from './services/message'
 import type { TakeoutEventFromCore, TakeoutEventToCore } from './services/takeout'
 
 import { useLogger } from '@tg-search/common'
