+++ packages/core/src/v2/event-handler.ts
@@ -9,9 +9,9 @@ import { createEmbeddingResolver } from './resolvers/embedding-resolver'
 import { createLinkResolver } from './resolvers/link-resolver'
 import { createUserResolver } from './resolvers/user-resolver'
 import { createConnectionService } from './services/connection'
-import { createDialogService } from './services/dialogs'
+import { createDialogService } from './services/dialog'
 import { createEntityService } from './services/entity'
-import { createMessageService } from './services/messages'
+import { createMessageService } from './services/message'
 import { createSessionService } from './services/session'
 import { createTakeoutService } from './services/takeout'
