+++ packages/core/src/v2/services/index.ts
@@ -1,5 +1,5 @@
 export type * from './connection'
-export type * from './dialogs'
-export type * from './messages'
+export type * from './dialog'
+export type * from './message'
 export type * from './session'
 export type * from './takeout'
